[PATCH] Increase PVR parameter buffer size
The default of 2 MB is far too small for the complex OpenGL rendering
used in the Android UI in ICS and up.  Instead, boost the param buffer
size to 12 MB, which improves performance.

XXX TODO: http://processors.wiki.ti.com/index.php/SGXDbg recommends 16
MB or larger, but this causes SurfaceFlinger to crash and burn.